requirements.txt
Flask
pyodbc
flask-jwt-extended
PyYAML
mysql-connector-python==9.0.0
authlib